Improve C consistency

+++ ext/cryptography/na_cl/na_cl.h
@@ -10,23 +10,23 @@
 #define RB_NACL_CHECK_STRING(str) \
     Check_Type((str), T_STRING)
 
-#define RB_NACL_CHECK_STRING_LEN(str, len)                                \
-    do {                                                                  \
-        RB_NACL_CHECK_STRING((str));                                      \
-        if (RSTRING_LEN((str)) != (len))                                  \
-            rb_raise(rb_eArgError, #str " must be %d bytes long", (len)); \
+#define RB_NACL_CHECK_STRING_LEN(str, len)                                  \
+    do {                                                                    \
+        RB_NACL_CHECK_STRING((str));                                        \
+        if ((unsigned long long)RSTRING_LEN((str)) != (len))                \
+            rb_raise(rb_eArgError, #str " must be %llu bytes long", (len)); \
     } while (0)
 
-#define RB_NACL_CALLOC(p, len, block) \
+#define RB_NACL_CALLOC(p, len, block)              \
     if (!((p) = calloc((len), sizeof((p))))) {     \
         block;                                     \
         rb_raise(rb_eNoMemError, "NoMemoryError"); \
     }
 
-#define RB_NACL_CFREE(p, len) \
-    if ((p)) {                \
-        memset(p, 0, len);    \
-        free(p);              \
+#define RB_NACL_CFREE(p, len)  \
+    if ((p)) {                 \
+        memset(p, 0, len);     \
+        free(p);               \
     }
 
 extern VALUE cNaClError;
